+++ zebra-network/src/peer_set/initialize.rs
@@ -265,6 +265,8 @@ where
         address_metrics,
         MinimumPeerVersion::new(latest_chain_tip, &config.network),
         None,
+        None,
+        None,
     );
     let peer_set = Buffer::new(BoxService::new(peer_set), constants::PEERSET_BUFFER_SIZE);
 
@@ -663,8 +665,12 @@ where
         + Clone,
     S::Future: Send + 'static,
 {
-    let mut recent_inbound_connections =
-        recent_by_ip::RecentByIp::new(None, Some(config.max_connections_per_ip));
+    let mut recent_inbound_connections = recent_by_ip::RecentByIp::new(
+        None,
+        Some(config.max_connections_per_ip),
+        Some(config.max_connections_per_subnet_v6),
+        Some(config.max_connections_per_subnet_v4),
+    );
 
     let mut active_inbound_connections = ActiveConnectionCounter::new_counter_with(
         config.peerset_inbound_connection_limit(),